+++ R/cloglogModel.R
@@ -13,6 +13,7 @@
 #' @importFrom maxLik maxLik
 #' @importFrom Matrix Matrix
 #' @importFrom survey svyrecvar
+#' @keywords internal
 #' @export
 # must be exported to be visible in c++ script, to consider any other option
 cloglog_model_nonprobsvy <- function(...) {